Move more symbols into Foundation
StackProperty, StackOffset, TooltipContent
