+++ hvega/src/Graphics/Vega/VegaLite/Core.hs
@@ -29,8 +29,6 @@ module Graphics.Vega.VegaLite.Core
        , BinProperty(..)
 
        , stack
-       , StackProperty(..)
-       , StackOffset(..)
 
        , calculateAs
 
@@ -66,7 +64,6 @@ module Graphics.Vega.VegaLite.Core
        , PointMarker(..)
        , LineMarker(..)
        , MarkErrorExtent(..)
-       , TooltipContent(..)
 
        , encoding
 
@@ -199,7 +196,6 @@ module Graphics.Vega.VegaLite.Core
        , boundsSpec
        , legendOrientLabel
        , compositionAlignmentSpec
-       , stackOffset
        , titleConfigSpec
        , autosizeProperty
        , paddingSpec
@@ -254,6 +250,9 @@ import Graphics.Vega.VegaLite.Foundation
   , OverlapStrategy
   , Side
   , Symbol
+  , StackProperty
+  , StackOffset
+  , TooltipContent(TTNone)
   , field_
   , order_
   , fontWeightSpec
@@ -272,6 +271,10 @@ import Graphics.Vega.VegaLite.Foundation
   , overlapStrategyLabel
   , sideLabel
   , symbolLabel
+  , stackPropertySpecSort
+  , stackPropertySpecOffset
+  , stackOffset
+  , ttContentLabel
   )
 import Graphics.Vega.VegaLite.Input
   ( Data
@@ -1192,57 +1195,6 @@ operationSpec Variance = "variance"
 operationSpec VarianceP = "variancep"
 
 
--- | How are stacks applied within a transform?
---
---   Prior to version @0.4.0.0@ the @StackProperty@ type was
---   what is now @StackOffset@.
-
-data StackProperty
-    = StOffset StackOffset
-      -- ^ Stack offset.
-      --
-      --   @since 0.4.0.0
-    | StSort [SortField]
-      -- ^ Ordering within a stack.
-      --
-      --   @since 0.4.0.0
-
-
--- | Describes the type of stacking to apply to a bar chart.
---
---   In @0.4.0.0@ this was renamed from @StackProperty@ to @StackOffset@,
---   but the constructor names have not changed.
---
-data StackOffset
-    = StZero
-      -- ^ Offset a stacked layout using a baseline at the foot of
-      --   the stack.
-    | StNormalize
-      -- ^ Rescale a stacked layout to use a common height while
-      --   preserving the relative size of stacked quantities.
-    | StCenter
-      -- ^ Offset a stacked layout using a central stack baseline.
-    | NoStack
-      -- ^ Do not stack marks, but create a layered plot.
-
-stackOffsetSpec :: StackOffset -> VLSpec
-stackOffsetSpec StZero = "zero"
-stackOffsetSpec StNormalize = "normalize"
-stackOffsetSpec StCenter = "center"
-stackOffsetSpec NoStack = A.Null
-
-stackOffset :: StackOffset -> LabelledSpec
-stackOffset so = "stack" .= stackOffsetSpec so
-
-
-stackPropertySpecOffset , stackPropertySpecSort:: StackProperty -> Maybe VLSpec
-stackPropertySpecOffset (StOffset op) = Just (stackOffsetSpec op)
-stackPropertySpecOffset _ = Nothing
-
-stackPropertySpecSort (StSort sfs) = Just (toJSON (map sortFieldSpec sfs))
-stackPropertySpecSort _ = Nothing
-
-
 {-|
 
 Apply a stack transform for positioning multiple values. This is an alternative
@@ -1255,7 +1207,7 @@ to specifying stacking directly when encoding position.
         []
         \"stack_count_Origin1\"
         \"stack_count_Origin2\"
-        [ 'StOffset' 'StNormalize', 'StSort' [ 'Graphics.Vega.VegaLite.WAscending' \"Origin\" ] ]
+        [ 'Graphics.Vega.VegaLite.StOffset' 'Graphics.Vega.VegaLite.StNormalize', 'Graphics.Vega.VegaLite.StSort' [ 'Graphics.Vega.VegaLite.WAscending' \"Origin\" ] ]
     . 'window'
         [ ( [ 'WAggregateOp' 'Min', 'WField' \"stack_count_Origin1\" ], \"x\" )
         , ( [ 'WAggregateOp' 'Max', 'WField' \"stack_count_Origin2\" ], \"x2\" )
@@ -1265,7 +1217,7 @@ to specifying stacking directly when encoding position.
         [ \"Origin\" ]
         \"y\"
         \"y2\"
-        [ 'StOffset' 'StNormalize', 'StSort' [ 'Graphics.Vega.VegaLite.WAscending' \"Cylinders\" ] ]
+        [ 'Graphics.Vega.VegaLite.StOffset' 'Graphics.Vega.VegaLite.StNormalize', 'Graphics.Vega.VegaLite.StSort' [ 'Graphics.Vega.VegaLite.WAscending' \"Cylinders\" ] ]
 @
 
 @since 0.4.0.0
@@ -1764,7 +1716,7 @@ data PositionChannel
       --           . 'position' 'Graphics.Vega.VegaLite.X' ['PName' \"week\", 'PmType' 'Graphics.Vega.VegaLite.Ordinal']
       --           . 'position' 'Graphics.Vega.VegaLite.Y' [ 'PName' \"takings\"
       --                        , 'PmType' 'Graphics.Vega.VegaLite.Quantitative'
-      --                        , 'PStack' 'StCenter'
+      --                        , 'PStack' 'Graphics.Vega.VegaLite.StCenter'
       --                        ]
       --           . 'color' ['MName' \"shop\", 'MmType' 'Graphics.Vega.VegaLite.Nominal']
       --   @
@@ -6235,35 +6187,3 @@ tooltips ::
   -> BuildLabelledSpecs
 tooltips tDefs ols =
   ("tooltip" .= toJSON (map (object . concatMap textChannelProperty) tDefs)) : ols
-
-
--- | This is used with 'MTooltip'.
---
---   @since 0.4.0.0
-
-data TooltipContent
-  = TTEncoding
-    -- ^ Tooltips are generated by the encoding (this is the default).
-    --
-    --   For example:
-    --
-    --   @'mark' 'Circle' ['MTooltip' 'TTEncoding']@
-  | TTData
-    -- ^ Tooltips are generated by all fields in the underlying data.
-    --
-    --   For example:
-    --
-    --   @'mark' 'Circle' ['MTooltip' 'TTData']@
-  | TTNone
-    -- ^ Disable tooltips.
-    --
-    --   For example:
-    --
-    --   @'mark' 'Circle' ['MTooltip' 'TTNone']@
-
-
--- Note that TTNone is special cased by markProperty
-ttContentLabel :: TooltipContent -> T.Text
-ttContentLabel TTEncoding = "encoding"
-ttContentLabel TTData = "data"
-ttContentLabel TTNone = "null"
