+++ hvega/src/Graphics/Vega/VegaLite/Foundation.hs
@@ -43,6 +43,11 @@ module Graphics.Vega.VegaLite.Foundation
 
        , Symbol(..)
 
+       , StackProperty(..)
+       , StackOffset(..)
+
+       , TooltipContent(..)
+
        -- not for external export
        , fontWeightSpec
        , measurementLabel
@@ -60,12 +65,17 @@ module Graphics.Vega.VegaLite.Foundation
        , overlapStrategyLabel
        , sideLabel
        , symbolLabel
+       , stackPropertySpecSort
+       , stackPropertySpecOffset
+       , stackOffset
+       , ttContentLabel
 
        , field_
        , order_
        )
     where
 
+import qualified Data.Aeson as A
 import qualified Data.Text as T
 
 import Data.Aeson ((.=), object, toJSON)
@@ -738,3 +748,86 @@ symbolLabel SymArrow = "arrow"
 symbolLabel SymTriangle = "triangle"
 symbolLabel SymWedge = "wedge"
 symbolLabel (SymPath svgPath) = svgPath
+
+
+-- | How are stacks applied within a transform?
+--
+--   Prior to version @0.4.0.0@ the @StackProperty@ type was
+--   what is now @StackOffset@.
+
+data StackProperty
+    = StOffset StackOffset
+      -- ^ Stack offset.
+      --
+      --   @since 0.4.0.0
+    | StSort [SortField]
+      -- ^ Ordering within a stack.
+      --
+      --   @since 0.4.0.0
+
+
+-- | Describes the type of stacking to apply to a bar chart.
+--
+--   In @0.4.0.0@ this was renamed from @StackProperty@ to @StackOffset@,
+--   but the constructor names have not changed.
+--
+data StackOffset
+    = StZero
+      -- ^ Offset a stacked layout using a baseline at the foot of
+      --   the stack.
+    | StNormalize
+      -- ^ Rescale a stacked layout to use a common height while
+      --   preserving the relative size of stacked quantities.
+    | StCenter
+      -- ^ Offset a stacked layout using a central stack baseline.
+    | NoStack
+      -- ^ Do not stack marks, but create a layered plot.
+
+stackOffsetSpec :: StackOffset -> VLSpec
+stackOffsetSpec StZero = "zero"
+stackOffsetSpec StNormalize = "normalize"
+stackOffsetSpec StCenter = "center"
+stackOffsetSpec NoStack = A.Null
+
+stackOffset :: StackOffset -> LabelledSpec
+stackOffset so = "stack" .= stackOffsetSpec so
+
+
+stackPropertySpecOffset , stackPropertySpecSort:: StackProperty -> Maybe VLSpec
+stackPropertySpecOffset (StOffset op) = Just (stackOffsetSpec op)
+stackPropertySpecOffset _ = Nothing
+
+stackPropertySpecSort (StSort sfs) = Just (toJSON (map sortFieldSpec sfs))
+stackPropertySpecSort _ = Nothing
+
+
+-- | This is used with 'Graphics.Vega.VegaLite.MTooltip'.
+--
+--   @since 0.4.0.0
+
+data TooltipContent
+  = TTEncoding
+    -- ^ Tooltips are generated by the encoding (this is the default).
+    --
+    --   For example:
+    --
+    --   @'Graphics.Vega.VegaLite.mark' 'Graphics.Vega.VegaLite.Circle' ['Graphics.Vega.VegaLite.MTooltip' 'TTEncoding']@
+  | TTData
+    -- ^ Tooltips are generated by all fields in the underlying data.
+    --
+    --   For example:
+    --
+    --   @'Graphics.Vega.VegaLite.mark' 'Graphics.Vega.VegaLite.Circle' ['Graphics.Vega.VegaLite.MTooltip' 'TTData']@
+  | TTNone
+    -- ^ Disable tooltips.
+    --
+    --   For example:
+    --
+    --   @'Graphics.Vega.VegaLite.mark' 'Graphics.Vega.VegaLite.Circle' ['Graphics.Vega.VegaLite.MTooltip' 'TTNone']@
+
+
+-- Note that TTNone is special cased by markProperty
+ttContentLabel :: TooltipContent -> T.Text
+ttContentLabel TTEncoding = "encoding"
+ttContentLabel TTData = "data"
+ttContentLabel TTNone = "null"
